[BOOK-67] refactor: 컴포넌트 스캔 범위 설정 및 스프링 어플리케이션 패키지 이동
