+++ apis/src/main/kotlin/org/yapp/apis/ApisApplication.kt
@@ -6,7 +6,14 @@ import org.springframework.boot.runApplication
 /**
  * Main application class for the apis module.
  */
-@SpringBootApplication
+@SpringBootApplication(
+    scanBasePackages = [
+        "org.yapp.apis",
+        "org.yapp.infra",
+        "org.yapp.domain",
+        "org.yapp.gateway"
+    ]
+)
 class ApisApplication
 
 fun main(args: Array<String>) {
